src/ml/main.py
import os
import logging
from typing import Dict, List, Optional
from fastapi import FastAPI, HTTPException, Security, Depends
from fastapi.security import OAuth2PasswordBearer
from pydantic import BaseModel
import uvicorn
from elasticsearch import Elasticsearch
import structlog
from prometheus_client import start_http_server, Counter, Histogram
import mlflow
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Configure structured logging
structlog.configure(
    processors=[
        structlog.processors.TimeStamper(fmt="iso"),
        structlog.processors.JSONRenderer()
    ]
)
logger = structlog.get_logger()

# Initialize FastAPI app
app = FastAPI(
    title="WazuhAI ML Engine",
    description="Machine Learning Engine for Wazuh SIEM",
    version="1.0.0"
)

# Initialize metrics
PREDICTION_COUNTER = Counter('ml_predictions_total', 'Total number of predictions made')
PREDICTION_LATENCY = Histogram('ml_prediction_latency_seconds', 'Prediction latency in seconds')

# Initialize Elasticsearch client
es = Elasticsearch(
    [os.getenv('ELASTICSEARCH_URL', 'http://elasticsearch:9200')],
    basic_auth=(os.getenv('ELASTICSEARCH_USERNAME'), os.getenv('ELASTICSEARCH_PASSWORD')),
    verify_certs=True
)

# Initialize MLflow
mlflow.set_tracking_uri(os.getenv('MLFLOW_TRACKING_URI', 'http://localhost:5000'))

# Security
oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")

class PredictionRequest(BaseModel):
    data: Dict
    model_version: Optional[str] = None

class PredictionResponse(BaseModel):
    prediction: Dict
    confidence: float
    model_version: str

@app.on_event("startup")
async def startup_event():
    """Initialize services on startup."""
    logger.info("starting_ml_engine")
    # Start Prometheus metrics server
    start_http_server(8000)
    # Load initial models
    load_models()

def load_models():
    """Load ML models from storage."""
    try:
        # Implementation for loading models
        logger.info("models_loaded_successfully")
    except Exception as e:
        logger.error("model_loading_failed", error=str(e))
        raise

@app.post("/predict", response_model=PredictionResponse)
async def predict(
    request: PredictionRequest,
    token: str = Depends(oauth2_scheme)
):
    """Make predictions using the ML model."""
    try:
        with PREDICTION_LATENCY.time():
            # Implementation for making predictions
            prediction = {"result": "sample_prediction"}
            confidence = 0.95
            model_version = "1.0.0"
            
            PREDICTION_COUNTER.inc()
            
            return PredictionResponse(
                prediction=prediction,
                confidence=confidence,
                model_version=model_version
            )
    except Exception as e:
        logger.error("prediction_failed", error=str(e))
        raise HTTPException(status_code=500, detail=str(e))

@app.get("/health")
async def health_check():
    """Health check endpoint."""
    return {"status": "healthy"}

@app.get("/models")
async def list_models(token: str = Depends(oauth2_scheme)):
    """List available ML models."""
    try:
        # Implementation for listing models
        return {"models": ["model1", "model2"]}
    except Exception as e:
        logger.error("model_listing_failed", error=str(e))
        raise HTTPException(status_code=500, detail=str(e))

if __name__ == "__main__":
    uvicorn.run(
        "main:app",
        host="0.0.0.0",
        port=5000,
        reload=True,
        ssl_keyfile=os.getenv('TLS_KEY'),
        ssl_certfile=os.getenv('TLS_CERT')
    ) 