src/ml/requirements.txt
# Core ML Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
torch==2.0.1
transformers==4.31.0

# Data Processing
elasticsearch==7.17.9
python-dateutil==2.8.2
pytz==2023.3

# API and Web
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6

# Security
cryptography==41.0.3
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.4

# Monitoring and Logging
prometheus-client==0.17.1
python-json-logger==2.0.7
structlog==23.1.0

# Testing
pytest==7.4.2
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Development Tools
black==23.7.0
flake8==6.1.0
mypy==1.5.1
isort==5.12.0

# Model Management
mlflow==2.7.1
boto3==1.28.23
azure-storage-blob==12.17.0
google-cloud-storage==2.10.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# Performance
numba==0.57.1
dask==2023.7.1
ray==2.7.0 